code style refactoring

+++ Heedbook/HBLib/Utils/SftpClient.cs
@@ -8,8 +8,8 @@ namespace HBLib.Utils
 {
     public class SftpClient : IDisposable
     {
-        private readonly SftpSettings _sftpSettings;
         private readonly Renci.SshNet.SftpClient _client;
+        private readonly SftpSettings _sftpSettings;
 
         public SftpClient(SftpSettings sftpSettings)
         {
@@ -18,19 +18,22 @@ namespace HBLib.Utils
             _sftpSettings = sftpSettings;
         }
 
+        public void Dispose()
+        {
+            _client.Dispose();
+        }
+
         private async Task ConnectToSftpAsync()
         {
             if (!_client.IsConnected)
-            {
-                await Task.Run(() => _client.Connect()).ContinueWith((t) =>
+                await Task.Run(() => _client.Connect()).ContinueWith(t =>
                 {
                     _client.ChangeDirectory(_sftpSettings.DestinationPath);
                 });
-            }
         }
 
         /// <summary>
-        /// Upload file to remote sftp server. 
+        ///     Upload file to remote sftp server.
         /// </summary>
         /// <param name="localPath"></param>
         /// <param name="remotePath"></param>
@@ -48,7 +51,7 @@ namespace HBLib.Utils
         }
 
         /// <summary>
-        /// Upload as memory stream to sftp server
+        ///     Upload as memory stream to sftp server
         /// </summary>
         /// <param name="stream"></param>
         /// <param name="path"></param>
@@ -63,7 +66,7 @@ namespace HBLib.Utils
         }
 
         /// <summary>
-        /// Downloads all files from specified directory. It's downloaded files to memory.
+        ///     Downloads all files from specified directory. It's downloaded files to memory.
         /// </summary>
         /// <param name="directory"></param>
         /// <returns>Returns ConcurrentDictionary String, MemoryStream where string is filename, Memory stream is file</returns>
@@ -76,7 +79,7 @@ namespace HBLib.Utils
             var tasks = files.Select(file =>
             {
                 var name = file.Name;
-                return new Task( () =>
+                return new Task(() =>
                 {
                     var ms = new MemoryStream();
                     _client.DownloadFile(directory + name, ms);
@@ -88,7 +91,7 @@ namespace HBLib.Utils
         }
 
         /// <summary>
-        /// Download one file from sftp as memory stream
+        ///     Download one file from sftp as memory stream
         /// </summary>
         /// <param name="path">Specific remote path of file. {folder}/{file}</param>
         /// <returns></returns>
@@ -101,7 +104,7 @@ namespace HBLib.Utils
         }
 
         /// <summary>
-        /// Download file to local disk.
+        ///     Download file to local disk.
         /// </summary>
         /// <param name="remotePath"></param>
         /// <returns></returns>
@@ -112,7 +115,7 @@ namespace HBLib.Utils
             var filename = remotePath.Split('/').Last();
 
             Console.WriteLine(localPath == null);
-            localPath = (localPath == null) ? localPath = _sftpSettings.DownloadPath + filename : localPath + filename;
+            localPath = localPath == null ? localPath = _sftpSettings.DownloadPath + filename : localPath + filename;
             Console.WriteLine($"{localPath}, {remotePath}");
             using (var fs = File.OpenWrite(localPath))
             {
@@ -123,7 +126,7 @@ namespace HBLib.Utils
         }
 
         /// <summary>
-        /// Check file exists on server
+        ///     Check file exists on server
         /// </summary>
         /// <param name="path">Specified directory + filename</param>
         /// <returns></returns>
@@ -134,22 +137,14 @@ namespace HBLib.Utils
         }
 
         /// <summary>
-        /// Delete file from server
+        ///     Delete file from server
         /// </summary>
         /// <param name="path">Specified directory + filename</param>
         /// <returns></returns>
         public async Task DeleteFileIfExistsAsync(String path)
         {
             await ConnectToSftpAsync();
-            if (_client.Exists(path))
-            {
-                await Task.Run(() => _client.DeleteFile(path));
-            }
-        }
-
-        public void Dispose()
-        {
-            _client.Dispose();
+            if (_client.Exists(path)) await Task.Run(() => _client.DeleteFile(path));
         }
     }
 }
